Add files via upload
This certification as a Digitalization Agent (Agente del Cambio) in Spain focused on guiding SMEs through their digital transformation processes, with a specialization in cybersecurity. The program combined 375 hours of training and mentoring, covering strategic business management, innovation, customer process optimization, and enabling technologies for digital transformation. It also included the development of soft skills and the creation of a tailored digital transformation plan for SMEs, preparing professionals to drive secure and efficient digital adoption.